requirements.txt
PyQt5
pytesseract
pyautogui
Pillow
schedule